+++ graft/coreth/peer/network_test.go
@@ -30,6 +30,10 @@ import (
 	"github.com/ava-labs/avalanchego/version"
 )
 
+const (
+	codecVersion uint16 = 0
+)
+
 var (
 	defaultPeerVersion = &version.Application{
 		Major: 1,
@@ -46,9 +50,7 @@ var (
 	_ message.RequestHandler = &HelloGreetingRequestHandler{}
 	_ message.RequestHandler = &testRequestHandler{}
 
-	_ common.AppSender      = testAppSender{}
-	_ message.GossipMessage = HelloGossip{}
-	_ message.GossipHandler = &testGossipHandler{}
+	_ common.AppSender = testAppSender{}
 
 	_ p2p.Handler = &testSDKHandler{}
 )
@@ -393,7 +395,7 @@ func TestRequestMinVersion(t *testing.T) {
 			go func() {
 				time.Sleep(200 * time.Millisecond)
 				atomic.AddUint32(&callNum, 1)
-				responseBytes, err := codecManager.Marshal(message.Version, TestMessage{Message: "this is a response"})
+				responseBytes, err := codecManager.Marshal(codecVersion, TestMessage{Message: "this is a response"})
 				if err != nil {
 					panic(err)
 				}
@@ -503,7 +505,6 @@ func TestHandleInvalidMessages(t *testing.T) {
 	p2pNetwork, err := p2p.NewNetwork(logging.NoLog{}, sender, prometheus.NewRegistry(), "")
 	require.NoError(t, err)
 	clientNetwork := NewNetwork(p2pNetwork, sender, codecManager, ids.EmptyNodeID, 1)
-	clientNetwork.SetGossipHandler(message.NoopMempoolGossipHandler{})
 	clientNetwork.SetRequestHandler(&testRequestHandler{})
 
 	assert.NoError(t, clientNetwork.Connected(context.Background(), nodeID, defaultPeerVersion))
@@ -511,7 +512,8 @@ func TestHandleInvalidMessages(t *testing.T) {
 	defer clientNetwork.Shutdown()
 
 	// Ensure a valid gossip message sent as any App specific message type does not trigger a fatal error
-	gossipMsg, err := buildGossip(codecManager, HelloGossip{Msg: "hello there!"})
+	marshaller := helloGossipMarshaller{codec: codecManager}
+	gossipMsg, err := marshaller.MarshalGossip(&HelloGossip{Msg: "hello there!"})
 	assert.NoError(t, err)
 
 	// Ensure a valid request message sent as any App specific message type does not trigger a fatal error
@@ -552,7 +554,6 @@ func TestNetworkPropagatesRequestHandlerError(t *testing.T) {
 	p2pNetwork, err := p2p.NewNetwork(logging.NoLog{}, nil, prometheus.NewRegistry(), "")
 	require.NoError(t, err)
 	clientNetwork := NewNetwork(p2pNetwork, sender, codecManager, ids.EmptyNodeID, 1)
-	clientNetwork.SetGossipHandler(message.NoopMempoolGossipHandler{})
 	clientNetwork.SetRequestHandler(&testRequestHandler{err: errors.New("fail")}) // Return an error from the request handler
 
 	assert.NoError(t, clientNetwork.Connected(context.Background(), nodeID, defaultPeerVersion))
@@ -615,18 +616,14 @@ func buildCodec(t *testing.T, types ...interface{}) codec.Manager {
 	for _, typ := range types {
 		assert.NoError(t, c.RegisterType(typ))
 	}
-	assert.NoError(t, codecManager.RegisterCodec(message.Version, c))
+	assert.NoError(t, codecManager.RegisterCodec(codecVersion, c))
 	return codecManager
 }
 
 // marshalStruct is a helper method used to marshal an object as `interface{}`
 // so that the codec is able to include the TypeID in the resulting bytes
 func marshalStruct(codec codec.Manager, obj interface{}) ([]byte, error) {
-	return codec.Marshal(message.Version, &obj)
-}
-
-func buildGossip(codec codec.Manager, msg message.GossipMessage) ([]byte, error) {
-	return codec.Marshal(message.Version, &msg)
+	return codec.Marshal(codecVersion, &obj)
 }
 
 type testAppSender struct {
@@ -696,11 +693,11 @@ type HelloGreetingRequestHandler struct {
 }
 
 func (h *HelloGreetingRequestHandler) HandleHelloRequest(ctx context.Context, nodeID ids.NodeID, requestID uint32, request *HelloRequest) ([]byte, error) {
-	return h.codec.Marshal(message.Version, HelloResponse{Response: "Hi"})
+	return h.codec.Marshal(codecVersion, HelloResponse{Response: "Hi"})
 }
 
 func (h *HelloGreetingRequestHandler) HandleGreetingRequest(ctx context.Context, nodeID ids.NodeID, requestID uint32, request *GreetingRequest) ([]byte, error) {
-	return h.codec.Marshal(message.Version, GreetingResponse{Greet: "Hey there"})
+	return h.codec.Marshal(codecVersion, GreetingResponse{Greet: "Hey there"})
 }
 
 type TestMessage struct {
@@ -719,34 +716,22 @@ type HelloGossip struct {
 	Msg string `serialize:"true"`
 }
 
-func (h HelloGossip) Handle(handler message.GossipHandler, nodeID ids.NodeID) error {
-	return handler.HandleEthTxs(nodeID, message.EthTxsGossip{})
+func (tx *HelloGossip) GossipID() ids.ID {
+	return ids.FromStringOrPanic(tx.Msg)
 }
 
-func (h HelloGossip) String() string {
-	return fmt.Sprintf("HelloGossip(%s)", h.Msg)
+type helloGossipMarshaller struct {
+	codec codec.Manager
 }
 
-func (h HelloGossip) Bytes() []byte {
-	// no op
-	return nil
+func (g helloGossipMarshaller) MarshalGossip(tx *HelloGossip) ([]byte, error) {
+	return g.codec.Marshal(0, tx)
 }
 
-type testGossipHandler struct {
-	received bool
-	nodeID   ids.NodeID
-}
-
-func (t *testGossipHandler) HandleAtomicTx(nodeID ids.NodeID, msg message.AtomicTxGossip) error {
-	t.received = true
-	t.nodeID = nodeID
-	return nil
-}
-
-func (t *testGossipHandler) HandleEthTxs(nodeID ids.NodeID, msg message.EthTxsGossip) error {
-	t.received = true
-	t.nodeID = nodeID
-	return nil
+func (g helloGossipMarshaller) UnmarshalGossip(bytes []byte) (*HelloGossip, error) {
+	h := &HelloGossip{}
+	_, err := g.codec.Unmarshal(bytes, h)
+	return h, err
 }
 
 type testRequestHandler struct {