+++ graft/coreth/plugin/evm/block.go
@@ -225,8 +225,9 @@ func (b *Block) handlePrecompileAccept(rules params.Rules) error {
 func (b *Block) Reject(context.Context) error {
 	log.Debug(fmt.Sprintf("Rejecting block %s (%s) at height %d", b.ID().Hex(), b.ID(), b.Height()))
 	for _, tx := range b.atomicTxs {
+		// Re-issue the transaction in the mempool, continue even if it fails
 		b.vm.mempool.RemoveTx(tx)
-		if err := b.vm.mempool.AddTx(tx); err != nil {
+		if err := b.vm.mempool.AddRemoteTx(tx); err != nil {
 			log.Debug("Failed to re-issue transaction in rejected block", "txID", tx.ID(), "err", err)
 		}
 	}