+++ graft/coreth/plugin/evm/gossip.go
@@ -1,13 +1,15 @@
 // Copyright (C) 2019-2023, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
+// TODO: move to network
+
 package evm
 
 import (
 	"context"
 	"fmt"
 	"sync"
-	syncatomic "sync/atomic"
+	"sync/atomic"
 	"time"
 
 	ethcommon "github.com/ethereum/go-ethereum/common"
@@ -24,7 +26,7 @@ import (
 	"github.com/ava-labs/coreth/core/txpool"
 	"github.com/ava-labs/coreth/core/types"
 	"github.com/ava-labs/coreth/eth"
-	"github.com/ava-labs/coreth/plugin/evm/atomic"
+	"github.com/ava-labs/coreth/plugin/evm/config"
 )
 
 const pendingTxsBuffer = 10
@@ -32,11 +34,9 @@ const pendingTxsBuffer = 10
 var (
 	_ p2p.Handler = (*txGossipHandler)(nil)
 
-	_ gossip.Gossipable                  = (*GossipEthTx)(nil)
-	_ gossip.Gossipable                  = (*GossipAtomicTx)(nil)
-	_ gossip.Marshaller[*GossipAtomicTx] = (*GossipAtomicTxMarshaller)(nil)
-	_ gossip.Marshaller[*GossipEthTx]    = (*GossipEthTxMarshaller)(nil)
-	_ gossip.Set[*GossipEthTx]           = (*GossipEthTxPool)(nil)
+	_ gossip.Gossipable               = (*GossipEthTx)(nil)
+	_ gossip.Marshaller[*GossipEthTx] = (*GossipEthTxMarshaller)(nil)
+	_ gossip.Set[*GossipEthTx]        = (*GossipEthTxPool)(nil)
 
 	_ eth.PushGossiper = (*EthPushGossiper)(nil)
 )
@@ -91,29 +91,14 @@ func (t txGossipHandler) AppRequest(ctx context.Context, nodeID ids.NodeID, dead
 	return t.appRequestHandler.AppRequest(ctx, nodeID, deadline, requestBytes)
 }
 
-type GossipAtomicTxMarshaller struct{}
-
-func (g GossipAtomicTxMarshaller) MarshalGossip(tx *GossipAtomicTx) ([]byte, error) {
-	return tx.Tx.SignedBytes(), nil
-}
-
-func (g GossipAtomicTxMarshaller) UnmarshalGossip(bytes []byte) (*GossipAtomicTx, error) {
-	tx, err := atomic.ExtractAtomicTx(bytes, atomic.Codec)
-	return &GossipAtomicTx{
-		Tx: tx,
-	}, err
-}
-
-type GossipAtomicTx struct {
-	Tx *atomic.Tx
-}
-
-func (tx *GossipAtomicTx) GossipID() ids.ID {
-	return tx.Tx.ID()
-}
-
 func NewGossipEthTxPool(mempool *txpool.TxPool, registerer prometheus.Registerer) (*GossipEthTxPool, error) {
-	bloom, err := gossip.NewBloomFilter(registerer, "eth_tx_bloom_filter", txGossipBloomMinTargetElements, txGossipBloomTargetFalsePositiveRate, txGossipBloomResetFalsePositiveRate)
+	bloom, err := gossip.NewBloomFilter(
+		registerer,
+		"eth_tx_bloom_filter",
+		config.TxGossipBloomMinTargetElements,
+		config.TxGossipBloomTargetFalsePositiveRate,
+		config.TxGossipBloomResetFalsePositiveRate,
+	)
 	if err != nil {
 		return nil, fmt.Errorf("failed to initialize bloom filter: %w", err)
 	}
@@ -134,7 +119,7 @@ type GossipEthTxPool struct {
 
 	// subscribed is set to true when the gossip subscription is active
 	// mostly used for testing
-	subscribed syncatomic.Bool
+	subscribed atomic.Bool
 }
 
 // IsSubscribed returns whether or not the gossip subscription is active.
@@ -161,7 +146,7 @@ func (g *GossipEthTxPool) Subscribe(ctx context.Context) {
 			return
 		case pendingTxs := <-g.pendingTxs:
 			g.lock.Lock()
-			optimalElements := (g.mempool.PendingSize(txpool.PendingFilter{}) + len(pendingTxs.Txs)) * txGossipBloomChurnMultiplier
+			optimalElements := (g.mempool.PendingSize(txpool.PendingFilter{}) + len(pendingTxs.Txs)) * config.TxGossipBloomChurnMultiplier
 			for _, pendingTx := range pendingTxs.Txs {
 				tx := &GossipEthTx{Tx: pendingTx}
 				g.bloom.Add(tx)